Quit button implemented. little bit of janky flair for making it work in the editor.
changed the name and moved a few things around, too. clean.

As an aside, you may note my other repos don't have nearly as many commits. this is mostly because they're for myself, and any issues that a lack of consistent and well documented commits may cause is future me's problem, and I hate that guy.

For cases where I'm working on a team, it's not just future me's problem, so I can't really morally justify getting away with that. I've had to fix enough repos already.